+++ core/src/main/java/com/leader/core/server/model/DynamicMessageFactory.java
@@ -6,7 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class DynamicMessageFactory {
-	private static HashMap<String, Integer> descriptors = new HashMap<String, Integer>();
+	private static HashMap<String, Short> descriptors = new HashMap<String, Short>();
 	private static Logger log = LoggerFactory.getLogger(DynamicMessageFactory.class);
 
 	/**
@@ -15,14 +15,14 @@ public class DynamicMessageFactory {
 	 * @param descriptor
 	 * @return
 	 */
-	public static int getDescriptor(String descriptor) {
+	public static short getDescriptor(String descriptor) {
 		if (!descriptors.containsKey(descriptor)) {
 			log.error("未找到对应的消息!");
 		}
 		return descriptors.get(descriptor);
 	}
 
-	public static void put(String descriptor, int type) {
+	public static void put(String descriptor, short type) {
 		descriptors.put(descriptor, type);
 	}
